+++ android/api-tests/src/test/java/com/revenuecat/apitests/kotlin/PaywallApiTests.kt
@@ -5,13 +5,11 @@ import com.revenuecat.purchases.Offering
 import com.revenuecat.purchases.hybridcommon.ui.PaywallResultListener
 import com.revenuecat.purchases.hybridcommon.ui.SingleFirePaywallResultListener
 import com.revenuecat.purchases.hybridcommon.ui.presentPaywallFromFragment
-import com.revenuecat.purchases.ui.revenuecatui.ExperimentalPreviewRevenueCatUIPurchasesAPI
 import com.revenuecat.purchases.ui.revenuecatui.activity.PaywallResult
 
 @Suppress("unused", "UNUSED_VARIABLE", "EmptyFunctionBlock")
 private class PaywallApiTests {
 
-    @OptIn(ExperimentalPreviewRevenueCatUIPurchasesAPI::class)
     fun checkPaywallResultListener() {
         val listener: PaywallResultListener = object : PaywallResultListener {
             override fun onPaywallResult(paywallResult: PaywallResult) {